requirements.txt
asgiref==3.3.1
Django==3.1.5
django-db-constraints==0.3.0
djangorestframework==3.12.2
psycopg2-binary==2.8.6
pytz==2020.5
selenium==3.141.0
sqlparse==0.4.1
urllib3==1.26.2

django-filter~=2.4.0
PyJWT~=2.0.1
channels~=3.0.3